refactor(rmcp-client): move program_resolver to separate module file (#3828)
Per review feedback, extracting program_resolver module to its own file
for better code organization and maintainability.

- Move program_resolver mod to src/program_resolver.rs
- Reorder Unix/Windows functions (Unix first as simpler)
- Move related tests to the new module file
- Add module declaration to lib.rs
